Update for DECIMAL changes

+++ src/NumpyDotNet/NumpyDotNet/function_base.cs
@@ -506,7 +506,7 @@ namespace NumpyDotNet
 
             // handle special case of a 1-element condition but
             // a multi-element choice
-            if (S.size == 1 || np.max(asanyarray(S.shape.iDims)) == 1)
+            if (S.size == 1 || (double)np.max(asanyarray(S.shape.iDims)) == 1)
             {
                 pfac = asarray(1);
                 for (int k = 0; k < n2 + 1; k++)
